Created All Tables
Note:

I have implemented it in the same way. The code is repetitive. I have tried to make a generic code (abstract, inheritance...) and it took too long. As this is a POC and we are under time limitation I will leave it as is for now and get back to it if and when we have time.