OOP Change - Controller
Same change as the revious commit just for the Controllers:

Created a base controller class to inherit from. There are functions that all controllers has:
getAll
create

So for maintainability and reusability I have created a base_controller to extend. In addition, I have bound "this(controller class this)" context to ensure 'this' is correct.